+++ GimbalFirmware/Headers/gopro/gopro_hero4.h
@@ -15,10 +15,10 @@
 
 // Hero 4 related state
 typedef struct {
-    uint16_t camera_proto_version[GP_H4_PROTO_NUM_BYTES];   // ZZ_REQ_SEND_PROTO_VERSION
-    uint16_t camera_fw_version[16];                         // non-null-terminated string
-    uint16_t channel_id;
-    uint16_t handshake_step;                                // GP_H4_HANDSHAKE_STEPS
+    uint8_t camera_proto_version[GP_H4_PROTO_NUM_BYTES];    // ZZ_REQ_SEND_PROTO_VERSION
+    uint8_t camera_fw_version[16];                          // non-null-terminated string
+    uint8_t channel_id;
+    uint8_t handshake_step;                                 // GP_H4_HANDSHAKE_STEPS
     bool pending_recording_state;   // latched based on response from camera
 } gp_h4_t;
 
@@ -26,55 +26,55 @@ typedef struct {
 
 // Request (CMD) General Format
 typedef struct {
-    uint16_t len;   // size of this packet (not including 'len')
-    uint16_t l1;    // first letter of 2-letter cmd
-    uint16_t l2;    // second letter of 2-letter cmd
-    uint16_t tid;   // transaction id (for future use, keep as 0)
-    uint16_t tcb;   // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
-    uint16_t payload[GP_H4_MAX_PAYLOAD - 5];
+    uint8_t len;    // size of this packet (not including 'len')
+    uint8_t l1;     // first letter of 2-letter cmd
+    uint8_t l2;     // second letter of 2-letter cmd
+    uint8_t tid;    // transaction id (for future use, keep as 0)
+    uint8_t tcb;    // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
+    uint8_t payload[GP_H4_MAX_PAYLOAD - 5];
 } gp_h4_cmd_t;
 
 // Response (RSP) General Format
 typedef struct {
-    uint16_t len;       // size of this packet (not including 'len')
-    uint16_t ack;       // NACK will occur if format of packet is unrecognized
-    uint16_t reserved;  // keep as 0
-    uint16_t tid;       // transaction id (keep as 0)
-    uint16_t tcb;       // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
-    uint16_t payload[GP_H4_MAX_PACKET - 5];
+    uint8_t len;        // size of this packet (not including 'len')
+    uint8_t ack;        // NACK will occur if format of packet is unrecognized
+    uint8_t reserved;   // keep as 0
+    uint8_t tid;        // transaction id (keep as 0)
+    uint8_t tcb;        // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
+    uint8_t payload[GP_H4_MAX_PACKET - 5];
 } gp_h4_rsp_t;
 
 #define GP_H4_YY_CMD_HEADER_SIZE 10
 #define GP_H4_YY_CMD_MAX_PAYLOAD (GP_H4_MAX_PACKET - GP_H4_YY_CMD_HEADER_SIZE) // TODO: necessary? maybe combine with GP_H4_MAX_PAYLOAD somehow?
 typedef struct {
-    uint16_t len;       // size of this packet (not including 'len')
-    uint16_t l1;        // first letter of 2-letter cmd
-    uint16_t l2;        // second letter of 2-letter cmd
-    uint16_t tid;       // transaction id (keep as 0)
-    uint16_t tcb;       // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
-    uint16_t chan_id;   // used as a session id for all subsequent YY requests
-    uint16_t api_group;
-    uint16_t api_id;
-    uint16_t datalen1;
-    uint16_t datalen2;
-    uint16_t payload[GP_H4_YY_CMD_MAX_PAYLOAD];
+    uint8_t len;        // size of this packet (not including 'len')
+    uint8_t l1;         // first letter of 2-letter cmd
+    uint8_t l2;         // second letter of 2-letter cmd
+    uint8_t tid;        // transaction id (keep as 0)
+    uint8_t tcb;        // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
+    uint8_t chan_id;    // used as a session id for all subsequent YY requests
+    uint8_t api_group;
+    uint8_t api_id;
+    uint8_t datalen1;
+    uint8_t datalen2;
+    uint8_t payload[GP_H4_YY_CMD_MAX_PAYLOAD];
 } gp_h4_yy_cmd_t;
 
 #define GP_H4_YY_RSP_HEADER_SIZE 11
 #define GP_H4_YY_RSP_MAX_PAYLOAD (GP_H4_MAX_PACKET - GP_H4_YY_RSP_HEADER_SIZE) // TODO: necessary? maybe combine with GP_H4_MAX_PAYLOAD somehow?
 typedef struct {
-    uint16_t len;       // size of this packet (not including 'len')
-    uint16_t ack;       // NACK will occur if format of packet is unrecognized
-    uint16_t reserved;  // keep as 0
-    uint16_t tid;       // transaction id (keep as 0)
-    uint16_t tcb;       // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
-    uint16_t chan_id;   // used as a session id for all subsequent YY requests
-    uint16_t api_group;
-    uint16_t api_id;
-    uint16_t err_code;  // zero == success, non-zero == error
-    uint16_t datalen1;
-    uint16_t datalen2;
-    uint16_t payload[GP_H4_YY_RSP_MAX_PAYLOAD];
+    uint8_t len;        // size of this packet (not including 'len')
+    uint8_t ack;        // NACK will occur if format of packet is unrecognized
+    uint8_t reserved;   // keep as 0
+    uint8_t tid;        // transaction id (keep as 0)
+    uint8_t tcb;        // Transaction Control Block bit mask (see GP_H4_TCB_VALS)
+    uint8_t chan_id;    // used as a session id for all subsequent YY requests
+    uint8_t api_group;
+    uint8_t api_id;
+    uint8_t err_code;   // zero == success, non-zero == error
+    uint8_t datalen1;
+    uint8_t datalen2;
+    uint8_t payload[GP_H4_YY_RSP_MAX_PAYLOAD];
 } gp_h4_yy_rsp_t;
 
 typedef union {
@@ -84,7 +84,7 @@ typedef union {
     gp_h4_yy_cmd_t yy_cmd;
     gp_h4_yy_rsp_t yy_rsp;
 
-    uint16_t bytes[GP_H4_MAX_PACKET];
+    uint8_t bytes[GP_H4_MAX_PACKET];
 
 } gp_h4_pkt_t;
 
@@ -104,9 +104,9 @@ void gp_h4_init(gp_h4_t *h4);
 bool gp_h4_handshake_complete(const gp_h4_t *h4);
 bool gp_h4_finish_handshake(gp_h4_t *h4, gp_h4_pkt_t *p);
 bool gp_h4_on_txn_complete(gp_h4_t *h4, gp_h4_pkt_t *p);
-bool gp_h4_recognize_packet(const uint16_t *buf, uint16_t len);
+bool gp_h4_recognize_packet(const uint8_t *buf, uint16_t len);
 
-bool gp_h4_rx_data_is_valid(const uint16_t *buf, uint16_t len);
+bool gp_h4_rx_data_is_valid(const uint8_t *buf, uint16_t len);
 gp_h4_err_t gp_h4_handle_rx(gp_h4_t *h4, const gp_h4_pkt_t *p, gp_h4_pkt_t *rsp);
 
 bool gp_h4_produce_get_request(gp_h4_t *h4, uint8_t cmd_id, gp_h4_pkt_t *p);
